# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Liveku Laku - Flood to Priority Card
  version: 0.1.0
  description: |
    Single sync endpoint for prelim. Frontend batches one Window (10s)
    of comments, POSTs once, gets one Priority Card back. Frozen at build.

paths:
  /catalog/parse:
    post:
      summary: Parse a product link or share-text into Seller Catalog fields
      description: |
        Setup-time helper, called once when the seller adds a product. Stateless:
        parses in-request, stores nothing. Plain HTTP only, no scraping infra.
        Missing fields are listed in needs_manual so the FE form can ask the seller.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogParseRequest'
            example:
              url: "https://vt.tokopedia.com/t/ZS9BRhoHp6Vgc-mhs03/"
      responses:
        '200':
          description: Parsed product fields, possibly partial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogParseResponse'
              example:
                name: "POCO X8 Pro Series | POCO X8 Pro Max"
                price: null
                promo: null
                stock: null
                image_url: "https://p16-oec-sg.ibyteimg.com/tos-alisg-i-aphluv4xwc-sg/feeb630a9685455e95a08cfad86c05de~tplv-aphluv4xwc-resize-webp:260:260.webp"
                parse_source: link
                needs_manual: [price, promo, stock]

  /analyze:
    post:
      summary: Analyze one Window → one Priority Card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
            example:
              source: mock # mock | shopee | tiktok
              window_seconds: 10
              products:
                - name: "Kaos Oversize Hitam"
                  price: "99k"
                  promo: "free ongkir Jawa"
                  stock: 42
              comments:
                - user: budi_99
                  text: "kak harga berapa?"
                  platform: tiktok
                  ts: 1710000000
                - user: sari
                  text: "ongkir ke medan berapa kak?"
                  platform: shopee
                  ts: 1710000001
      responses:
        '200':
          description: Priority Card for that Window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
              example:
                total: 18
                window_seconds: 10
                clusters:
                  - label: harga
                    count: 9
                    share: 0.5
                  - label: ongkir
                    count: 5
                    share: 0.28
                top_cluster:
                  label: harga
                  label_id: Harga
                  count: 9
                  share: 0.5
                  samples:
                    - "kak harga berapa?"
                    - "spill harga dong"
                urgency: 82
                suggested_reply: "Harga Kaos Oversize 99k free ongkir Jawa kak - langsung checkout biar kebagian!"
                why_now: "9 dari 18 komen window ini nanya harga - jawab sekarang sebelum pindah live lain."
                tone: closing
                source: mock # mock | template | muse-spark-1.2-contributor

components:
  schemas:
    ProductIn:
      type: object
      required: [name]
      properties:
        name: {type: string}
        price: {type: string, nullable: true}
        promo: {type: string, nullable: true}
        stock: {type: integer, nullable: true}

    CatalogParseRequest:
      type: object
      required: []
      description: Provide url (product share link) or text (share-sheet blob). Empty input returns manual-entry defaults.
      properties:
        url:
          type: string
          nullable: true
          example: "https://id.shp.ee/i4befBuF"
        text:
          type: string
          nullable: true
          example: "Kaos Oversize Hitam Rp99.000 https://vt.tokopedia.com/t/ZS9BRhoHp6Vgc-mhs03/"

    CatalogParseResponse:
      type: object
      required: [parse_source, needs_manual]
      properties:
        name: {type: string, nullable: true}
        price: {type: string, nullable: true, description: normalized string like "99k" or "Rp99.000"}
        promo: {type: string, nullable: true}
        stock: {type: integer, nullable: true}
        image_url: {type: string, nullable: true}
        parse_source:
          type: string
          enum: [link, text, manual]
          description: how the fields were obtained; manual means nothing parsed, seller fills the form
        needs_manual:
          type: array
          items: {type: string, enum: [name, price, promo, stock]}

    AnalyzeRequest:
      type: object
      required: []
      properties:
        source:
          type: string
          enum: [mock, shopee, tiktok]
          example: mock
        window_seconds:
          type: integer
          example: 10
        handle:
          type: string
          example: "@tokoku"
          description: TikTok handle for live fetch inside request when source is tiktok and comments is empty
        session_id:
          type: string
          example: "6236215"
          description: Shopee session id for live fetch inside request when source is shopee and comments is empty
        products:
          type: array
          maxItems: 20
          description: Seller Catalog held client-side and sent inline each call; grounds suggested_reply in real products
          items:
            $ref: '#/components/schemas/ProductIn'
        comments:
          type: array
          minItems: 0
          maxItems: 80
          items:
            $ref: '#/components/schemas/CommentIn'

    CommentIn:
      type: object
      required: [text]
      properties:
        user:
          type: string
          example: budi_99
        text:
          type: string
          example: "kak harga berapa?"
        platform:
          type: string
          enum: [tiktok, shopee, mock]
        ts:
          type: integer
          description: unix seconds, optional

    AnalyzeResponse:
      type: object
      required: [total, window_seconds, clusters, top_cluster, urgency, why_now]
      properties:
        total:
          type: integer
        window_seconds:
          type: integer
        clusters:
          type: array
          items:
            type: object
            properties:
              label: {type: string}
              count: {type: integer}
              share: {type: number}
        top_cluster:
          type: object
          properties:
            label: {type: string}
            label_id: {type: string}
            count: {type: integer}
            share: {type: number}
            samples:
              type: array
              items: {type: string}
        urgency: {type: integer, minimum: 0, maximum: 100}
        suggested_reply:
          type: string
          nullable: true
          description: Optional coaching the host may speak; the topic is the main output. Null when no grounded suggestion is available.
        why_now: {type: string}
        tone: {type: string, enum: [closing, reassure, inform, upsell]}
        source:
          type: string
          enum: [mock, template, muse-spark-1.2-contributor]
          description: which engine produced the coaching; template is the deterministic keyless fallback